+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - Added code metrics
     - Added training and inference sections to the docs
     - Decision tree rules method now outputs a string
+    - Added drop row and column methods to dataset interface
 
 - 0.0.16-beta
     - Radius Neighbors allows user-definable anomaly class